Modified & Add, Login component created & added suspense with fallback to LoginPage
